+++ src/main/java/core/mvc/param/parser/StringParser.java
@@ -1,4 +1,4 @@
-package core.mvc.param;
+package core.mvc.param.parser;
 
 public class StringParser implements Parser<String> {
     @Override
