+++ src/test/java/core/mvc/param/parser/TypeParserTest.java
@@ -1,4 +1,4 @@
-package core.mvc.param;
+package core.mvc.param.parser;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
